Major refactor:
Added classes for basic/C-style types (int8, char, etc.).
This improves the semantics of the types used by the parser,
as well as acts as an alert for issues in the packet definitions
or parser by allowing exceptions to be thrown if the type's parsed
value is out of the expected range.

Improved parsing logic down to a single recursive function by
making better use of the new types.

Added some example packet data for testing.